requirements.txt
Django==5.2.4
djangorestframework==3.15.2
pyTelegramBotAPI==4.15.4
requests==2.31.0
python-dotenv==1.0.0 